Replace glog with boost.log
 - Investigate alternative logging frameworks to glog: Boost.Log.
 - Implement in Shared/Logger.{h,cpp}
 - Support both file and console (std::clog) sinks.
 - Replaced #include <glog/logging.h> with #include "Shared/Logger.h"
 - Removed logging from Calcite destructor which caused problems on shutdown.
 - Move std::ostream operator<<() overloads into their namespaces to allow for ADL.
 - Print Boost version in cmake.
 - Don't log when __CUDACC__ is defined.
 - Move always_assert.h logic into Logger.h and delete file.
 - Move unreachable.h logic into Logger.h and delete file.
 - Fix casting in numerous places. Warnings were previously suppressed by glog.

+++ Fragmenter/SortedOrderFragmenter.cpp
@@ -214,7 +214,7 @@ void SortedOrderFragmenter::sortData(InsertData& insertDataStruct) {
   std::vector<size_t> indexes(insertDataStruct.numRows);
   std::iota(indexes.begin(), indexes.end(), 0);
   const auto dist = std::distance(insertDataStruct.columnIds.begin(), it);
-  CHECK_LT(dist, insertDataStruct.data.size());
+  CHECK_LT(static_cast<size_t>(dist), insertDataStruct.data.size());
   sortIndexes(physical_cd, indexes, insertDataStruct.data[dist]);
   // shuffle rows of all columns
   for (size_t i = 0; i < insertDataStruct.columnIds.size(); ++i) {
